tests now work
