+++ requirements.txt
@@ -1,5 +1,6 @@
 annotated-types==0.7.0
 anyio==4.4.0
+asyncio==3.4.3
 attrs==23.2.0
 bidict==0.23.1
 blinker==1.8.2
@@ -11,11 +12,13 @@ dnspython==2.6.1
 email_validator==2.2.0
 fastapi==0.111.0
 fastapi-cli==0.0.4
+-e git+https://github.com/AberdeenRSS/FlightManagementServer.git@2f0369fd5e56140f6d01ad5f41c24fb7226b2731#egg=flightmanagementserver
 h11==0.14.0
 httpcore==1.0.5
 httptools==0.6.1
 httpx==0.27.0
 idna==3.7
+iniconfig==2.0.0
 Jinja2==3.1.4
 jsonschema==4.22.0
 jsonschema-specifications==2023.12.1
@@ -25,6 +28,8 @@ mdurl==0.1.2
 motor==3.5.0
 netifaces==0.10.6
 orjson==3.10.6
+packaging==24.1
+pluggy==1.5.0
 pycparser==2.22
 pydantic==2.8.2
 pydantic-settings==2.3.4
@@ -32,15 +37,17 @@ pydantic_core==2.20.1
 Pygments==2.18.0
 PyJWT==2.8.0
 pymongo==4.8.0
+pytest==8.3.2
+pytest-asyncio==0.24.0
 python-dotenv==1.0.1
 python-engineio==4.9.1
 python-multipart==0.0.9
 python-socketio==5.11.3
+pytz==2024.2
 PyYAML==6.0.1
 referencing==0.35.1
 rich==13.7.1
 rpds-py==0.18.1
-setuptools==3.3
 shellingham==1.5.4
 simple-websocket==1.0.0
 sniffio==1.3.1
