+++ setup.py
@@ -1,0 +1,6 @@
+from setuptools import setup, find_packages
+
+setup(
+    name="flightmanagementserver",
+    packages=find_packages(),
+)
