+++ tests/running_tests.md
@@ -1,0 +1,11 @@
+# Running Tests
+
+### Setup
+
+Run a mongodb instance locally on port :27017
+
+Make sure you have both pem files in the root level of this repository.
+```
+private.pem
+public.pem
+```
